clean routes, started implementing new api design, implemented basic automation opeartions
