+++ app/mqtt_publish_automations.py
@@ -8,17 +8,17 @@ from contextlib import asynccontextmanager
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from paho.mqtt import client as mqtt_client
 
-from ..models import (
+from .models import (
     ContainerProperties,
     Topology,
     Pipeline,
     encode_pydantic_model,
 )
-from ..utils import calculate_uptime, read_bitswan_yaml
-from ..mqtt import mqtt_resource
+from .utils import calculate_uptime, read_bitswan_yaml
+from .mqtt import mqtt_resource
 
 
-async def retrieve_active_pres() -> Topology:
+async def retrieve_active_automations() -> Topology:
     client = docker.from_env()
     info = client.info()
 
@@ -67,7 +67,7 @@ async def retrieve_active_pres() -> Topology:
     return Topology(**topology)
 
 
-async def retrieve_inactive_pres() -> Topology:
+async def retrieve_inactive_automations() -> Topology:
     bs_home = os.environ.get("BITSWAN_BITSWAN_DIR", "/mnt/repo/pipeline")
     bs_yaml = read_bitswan_yaml(bs_home)
 
@@ -106,15 +106,15 @@ async def retrieve_inactive_pres() -> Topology:
     return Topology(**topology)
 
 
-async def publish_pres(client: mqtt_client.Client) -> Topology:
+async def publish_automations(client: mqtt_client.Client) -> Topology:
     topic = os.environ.get("MQTT_TOPIC", "bitswan/topology")
-    active = await retrieve_active_pres()
-    inactive = await retrieve_inactive_pres()
+    active = await retrieve_active_automations()
+    inactive = await retrieve_inactive_automations()
 
-    pres = inactive.topology.copy()
-    pres.update(active.topology)
+    automations = inactive.topology.copy()
+    automations.update(active.topology)
 
-    topology = Topology(topology=pres, display_style="list")
+    topology = Topology(topology=automations, display_style="list")
 
     client.publish(
         topic,
@@ -133,7 +133,7 @@ async def lifespan(app: FastAPI):
 
     if result:
         scheduler.add_job(
-            functools.partial(publish_pres, mqtt_resource.get_client()),
+            functools.partial(publish_automations, mqtt_resource.get_client()),
             trigger="interval",
             seconds=10,
         )
