+++ app/routes/automations.py
@@ -1,4 +1,5 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, UploadFile, File, HTTPException
+from fastapi.responses import JSONResponse
 from app.services.automation_service import AutomationService
 from app.dependencies import get_automation_service
 
@@ -8,45 +9,49 @@ router = APIRouter(prefix="/automations", tags=["automations"])
 async def get_automations(automation_service: AutomationService = Depends(get_automation_service)):
     return automation_service.get_automations()
 
-@router.post("/{deployment_id}")
-async def create_automation(deployment_id: str):
-    pass
-
-@router.delete("/{deployment_id}")
-async def delete_automation(deployment_id: str):
-    pass
+@router.post("/deploy")
+async def deploy_automations(automation_service: AutomationService = Depends(get_automation_service)):
+    return await automation_service.deploy_automations()
 
 @router.post("/{deployment_id}/deploy")
-async def deploy_automation(deployment_id: str):
-    pass
-
-@router.post("/deploy")
-async def deploy_automations():
-    pass
+async def deploy_automation(deployment_id: str, automation_service: AutomationService = Depends(get_automation_service)):
+    return automation_service.deploy_automation(deployment_id)
 
 @router.post("/{deployment_id}/start")
-async def start_automation(deployment_id: str):
-    pass
+async def start_automation(deployment_id: str, automation_service: AutomationService = Depends(get_automation_service)):
+    return automation_service.start_automation(deployment_id)
 
 @router.post("/{deployment_id}/stop")
-async def stop_automation(deployment_id: str):
-    pass
+async def stop_automation(deployment_id: str, automation_service: AutomationService = Depends(get_automation_service)):
+    return automation_service.stop_automation(deployment_id)
 
 @router.post("/{deployment_id}/restart")
-async def restart_automation(deployment_id: str):
-    pass
+async def restart_automation(deployment_id: str, automation_service: AutomationService = Depends(get_automation_service)):
+    return automation_service.restart_automation(deployment_id)
 
 @router.post("/{deployment_id}/activate")
-async def activate_automation(deployment_id: str):
-    pass
+async def activate_automation(deployment_id: str, automation_service: AutomationService = Depends(get_automation_service)):
+    return automation_service.activate_automation(deployment_id)
 
 @router.post("/{deployment_id}/deactivate")
-async def deactivate_automation(deployment_id: str):
-    pass
+async def deactivate_automation(deployment_id: str, automation_service: AutomationService = Depends(get_automation_service)):
+    return automation_service.deactivate_automation(deployment_id)
 
-@router.post("/{deployment_id}/logs")
-async def get_automation_logs(deployment_id: str, lines: int = 100):
-    pass
+@router.get("/{deployment_id}/logs")
+async def get_automation_logs(deployment_id: str, lines: int = 100, automation_service: AutomationService = Depends(get_automation_service)):
+    return automation_service.get_automation_logs(deployment_id, lines)
+
+@router.post("/{deployment_id}")
+async def create_automation(deployment_id: str, file: UploadFile = File(...), automation_service: AutomationService = Depends(get_automation_service)):
+    if file.filename.endswith(".zip"):
+        result = await automation_service.create_automation(deployment_id, file)
+        return JSONResponse(content=result)
+    else:
+        raise HTTPException(status_code=400, detail="File must be a ZIP archive")
+
+@router.delete("/{deployment_id}")
+async def delete_automation(deployment_id: str, automation_service: AutomationService = Depends(get_automation_service)):
+    return automation_service.delete_automation(deployment_id)
