+++ app/services/automation_service.py
@@ -1,15 +1,24 @@
 from datetime import datetime
 import os
+import shutil
+from tempfile import NamedTemporaryFile
+import zipfile
 import docker
+import yaml
 from app.models import DeployedPRE
-from app.utils import calculate_uptime, read_bitswan_yaml
-
+from app.utils import add_route_to_caddy, calculate_checksum, calculate_uptime, docker_compose_up, read_bitswan_yaml, read_pipeline_conf, update_git
+from fastapi import UploadFile, HTTPException
 
 class AutomationService:
+    def __init__(self):
+        self.bs_home = os.environ.get("BITSWAN_GITOPS_DIR", "/mnt/repo/pipeline")
+        self.bs_home_host = os.environ.get("BITSWAN_GITOPS_DIR_HOST", "/home/root/.config/bitswan/local-gitops/")
+        self.gitops_id = os.environ.get("BITSWAN_GITOPS_ID", "gitops-local")
+        self.docker_client = docker.from_env()
+
 
     def get_automations(self):
-        bs_home = os.environ.get("BITSWAN_GITOPS_DIR", "/mnt/repo/pipeline")
-        gitops_dir = os.path.join(bs_home, "gitops")
+        gitops_dir = os.path.join(self.bs_home, "gitops")
         bs_yaml = read_bitswan_yaml(gitops_dir)
 
         if not bs_yaml:
@@ -29,9 +38,8 @@ class AutomationService:
             for deployment_id in bs_yaml["deployments"]
         }
 
-        client = docker.from_env()
-        info = client.info()
-        containers: list[docker.models.containers.Container] = client.containers.list(
+        info = self.docker_client.info()
+        containers: list[docker.models.containers.Container] = self.docker_client.containers.list(
             filters={
                 "label": [
                     "space.bitswan.pipeline.protocol-version",
@@ -59,8 +67,58 @@ class AutomationService:
 
         return list(pres.values())
 
-    def create_automation(self, deployment_id: str):
-        pass
+    async def create_automation(self, deployment_id: str, file: UploadFile):
+        with NamedTemporaryFile(delete=False) as temp_file:
+            content = await file.read()
+            temp_file.write(content)
+
+            checksum = calculate_checksum(temp_file.name)
+            output_dir = f"{checksum}"
+            old_deploymend_checksum = None
+
+            try:
+                bitswan_path = os.path.join(self.bs_home, "gitops")
+                bitswan_path_host = os.path.join(self.bs_home_host, "gitops")
+
+                bitswan_yaml_path = os.path.join(bitswan_path, "bitswan.yaml")
+
+                output_dir = os.path.join(bitswan_path, output_dir)
+
+                os.makedirs(output_dir, exist_ok=True)
+                with zipfile.ZipFile(temp_file.name, "r") as zip_ref:
+                    zip_ref.extractall(output_dir)
+
+                # Update or create bitswan.yaml
+                data = read_bitswan_yaml(bitswan_path)
+
+                data = data or {"deployments": {}}
+                deployments = data["deployments"]  # should never raise KeyError
+
+                deployments[deployment_id] = deployments.get(deployment_id, {})
+                old_deploymend_checksum = deployments[deployment_id].get("checksum")
+                deployments[deployment_id]["checksum"] = checksum
+                deployments[deployment_id]["active"] = True
+
+                data["deployments"] = deployments
+                with open(bitswan_yaml_path, "w") as f:
+                    yaml.dump(data, f)
+
+                await update_git(bitswan_path, bitswan_path_host, deployment_id, checksum)
+
+                return {
+                    "message": "File processed successfully",
+                    "output_directory": output_dir,
+                    "checksum": checksum,
+                }
+            except Exception as e:
+                shutil.rmtree(output_dir, ignore_errors=True)
+                return {"error": f"Error processing file: {str(e)}"}
+            finally:
+                if old_deploymend_checksum:
+                    shutil.rmtree(
+                        os.path.join(bitswan_path, old_deploymend_checksum), ignore_errors=True
+                    )
+                os.unlink(temp_file.name)
 
     def delete_automation(self, deployment_id: str):
         pass
@@ -68,20 +126,215 @@ class AutomationService:
     def deploy_automation(self, deployment_id: str):
         pass
 
+    async def deploy_automations(self):
+        os.environ["COMPOSE_PROJECT_NAME"] = self.gitops_id
+
+        gitops_config = os.path.join(self.bs_home, "gitops")
+        gitops_config_host = os.path.join(self.bs_home_host, "gitops")
+        secrets_dir = os.path.join(self.bs_home, "secrets")
+
+        bs_yaml = read_bitswan_yaml(gitops_config)
+
+        if not bs_yaml:
+            raise HTTPException(status_code=500, detail="Error reading bitswan.yaml")
+
+        dc = {
+            "version": "3",
+            "services": {},
+        }
+        external_networks = {"bitswan_network"}
+        deployments = bs_yaml.get("deployments", {})
+        for deployment_id, conf in deployments.items():
+            conf = conf or {}
+            entry = {}
+
+            source = conf.get("source") or conf.get("checksum") or deployment_id
+            source_dir = os.path.join(gitops_config, source)
+
+            if not os.path.exists(source_dir):
+                raise HTTPException(status_code=500, detail=f"Deployment directory {source_dir} does not exist")
+            else:
+                pipeline_conf = read_pipeline_conf(source_dir)
+
+            entry["environment"] = {"DEPLOYMENT_ID": deployment_id}
+            entry["container_name"] = deployment_id
+            entry["restart"] = "always"
+            entry["labels"] = {
+                "gitops.deployment_id": deployment_id,
+            }
+            entry["image"] = "bitswan/pipeline-runtime-environment:latest"
+
+            network_mode = None
+            secret_groups = []
+            if pipeline_conf:
+                network_mode = pipeline_conf.get(
+                    "docker.compose", "network_mode", fallback=conf.get("network_mode")
+                )
+                secret_groups = pipeline_conf.get("secrets", "groups", fallback="").split(
+                    " "
+                )
+            for secret_group in secret_groups:
+                # Skip empty secret groups
+                if not secret_group:
+                    continue
+                if os.path.exists(secrets_dir):
+                    secret_env_file = os.path.join(secrets_dir, secret_group)
+                    if os.path.exists(secret_env_file):
+                        if not entry.get("env_file"):
+                            entry["env_file"] = []
+                        entry["env_file"].append(secret_env_file)
+
+            if not network_mode:
+                network_mode = conf.get("network_mode")
+
+            if network_mode:
+                entry["network_mode"] = network_mode
+            elif "networks" in conf:
+                entry["networks"] = conf["networks"].copy()
+            elif "default-networks" in bs_yaml:
+                entry["networks"] = bs_yaml["default-networks"].copy()
+            else:
+                entry["networks"] = ["bitswan_network"]
+            if entry.get("networks"):
+                external_networks.update(set(entry["networks"]))
+
+            passthroughs = ["volumes", "ports", "devices", "container_name"]
+            entry.update({p: conf[p] for p in passthroughs if p in conf})
+
+            deployment_dir = os.path.join(gitops_config_host, source)
+
+            if pipeline_conf:
+                entry["image"] = (
+                    pipeline_conf.get("deployment", "pre", fallback=entry.get("image"))
+                    or entry["image"]
+                )
+                expose = pipeline_conf.getboolean(
+                    "deployment", "expose", fallback=conf.get("expose")
+                )
+                port = pipeline_conf.get(
+                    "deployment", "port", fallback=conf.get("port", 8080)
+                )
+                if expose and port:
+                    result = add_route_to_caddy(deployment_id, port)
+                    if not result:
+                        raise HTTPException(status_code=500, detail="Error adding route to Caddy")
+
+            if "volumes" not in entry:
+                entry["volumes"] = []
+            entry["volumes"].append(f"{deployment_dir}:/opt/pipelines")
+
+            if conf.get("enabled", True):
+                dc["services"][deployment_id] = entry
+
+        dc["networks"] = {}
+        for network in external_networks:
+            dc["networks"][network] = {"external": True}
+        dc_yaml = yaml.dump(dc)
+
+        deployment_result = await docker_compose_up(gitops_config, dc_yaml, deployments)
+
+        for result in deployment_result.values():
+            if result["return_code"] != 0:
+                raise HTTPException(status_code=500, detail="Error deploying services")
+        return {
+            "message": "Deployed services successfully",
+            "deployments": list(deployments.keys()),
+            "result": deployment_result,
+        }
+
     def start_automation(self, deployment_id: str):
-        pass
+        containers = self.docker_client.containers.list(
+            all=True,  # Include stopped containers
+            filters={
+                "label": [
+                    "space.bitswan.pipeline.protocol-version",
+                    f"gitops.deployment_id={deployment_id}"
+                ]
+            }
+        )
+        
+        if not containers:
+            raise HTTPException(status_code=404, detail=f"No container found for deployment ID: {deployment_id}")
+        
+        # Restart the container
+        container = containers[0]
+        container.start()
+        
+        return {
+            "status": "success",
+            "message": f"Container for deployment {deployment_id} started successfully"
+        }
 
     def stop_automation(self, deployment_id: str):
-        pass
+        containers = self.docker_client.containers.list(
+            all=True,  # Include stopped containers
+            filters={
+                "label": [
+                    "space.bitswan.pipeline.protocol-version",
+                    f"gitops.deployment_id={deployment_id}"
+                ]
+            }
+        )
+        
+        if not containers:
+            raise HTTPException(status_code=404, detail=f"No container found for deployment ID: {deployment_id}")
+        
+        container = containers[0]
+        container.stop()
+        
+        return {
+            "status": "success",
+            "message": f"Container for deployment {deployment_id} stopped successfully"
+        }
 
     def restart_automation(self, deployment_id: str):
-        pass
+        containers = self.docker_client.containers.list(
+            all=True,  # Include stopped containers
+            filters={
+                "label": [
+                    "space.bitswan.pipeline.protocol-version",
+                    f"gitops.deployment_id={deployment_id}"
+                ]
+            }
+        )
+        
+        if not containers:
+            raise HTTPException(status_code=404, detail=f"No container found for deployment ID: {deployment_id}")
+        
+        # Restart the container
+        container = containers[0]
+        container.restart()
+        
+        return {
+            "status": "success",
+            "message": f"Container for deployment {deployment_id} restarted successfully"
+        }
 
     def activate_automation(self, deployment_id: str):
-        pass
+        gitops_dir = os.path.join(self.bs_home, "gitops")
 
     def deactivate_automation(self, deployment_id: str):
         pass
 
     def get_automation_logs(self, deployment_id: str, lines: int = 100):
-        pass
+        containers = self.docker_client.containers.list(
+            all=True,  # Include stopped containers
+            filters={
+                "label": [
+                    "space.bitswan.pipeline.protocol-version",
+                    f"gitops.deployment_id={deployment_id}"
+                ]
+            }
+        )
+
+        if not containers:
+            raise HTTPException(status_code=404, detail=f"No container found for deployment ID: {deployment_id}")
+        
+        container = containers[0]
+        logs = container.logs(tail=lines)
+        logs = logs.decode("utf-8")
+
+        return {
+            "status": "success",
+            "logs": logs.split("\n")
+        }
